+++ commands/src/main/java/org/wildfly/extras/creaper/commands/infinispan/cache/AbstractAddCache.java
@@ -11,10 +11,6 @@ abstract class AbstractAddCache implements OnlineCommand {
     private final String cacheContainer;
     private final String name;
 
-    // cache properties common to all cache types
-    private final String jndiName;
-    private final String module;
-    private final String start;
     private final Boolean statisticsEnabled;
 
     protected final Address address;
@@ -26,9 +22,6 @@ abstract class AbstractAddCache implements OnlineCommand {
 
         this.cacheContainer = builder.cacheContainer;
         this.name = builder.name;
-        this.jndiName = builder.jndiName;
-        this.module = builder.module;
-        this.start = builder.start;
         this.statisticsEnabled = builder.statisticsEnabled;
 
         this.address = Address.subsystem("infinispan")
@@ -38,15 +31,7 @@ abstract class AbstractAddCache implements OnlineCommand {
 
     @Override
     public final void apply(OnlineCommandContext ctx) throws Exception {
-        if (ctx.version.greaterThanOrEqualTo(ServerVersion.VERSION_20_0_0)) {
-            throw new AssertionError("This command is not compatible with WildFly 27 and above, "
-                    + "see https://github.com/wildfly-extras/creaper/issues/218.");
-        }
-
         Values values = Values.empty()
-                .andOptional("jndi-name", jndiName)
-                .andOptional("module", module)
-                .andOptional("start", start)
                 .andOptional("statistics-enabled", statisticsEnabled);
         values = addValuesSpecificForCacheType(values, ctx.version);
 
@@ -59,9 +44,6 @@ abstract class AbstractAddCache implements OnlineCommand {
     abstract static class Builder<THIS extends Builder> {
         protected String name;
         protected String cacheContainer;
-        protected String jndiName;
-        protected String module;
-        protected String start;
         protected Boolean statisticsEnabled;
 
         protected Builder(String name) {
@@ -73,21 +55,6 @@ abstract class AbstractAddCache implements OnlineCommand {
 
         public final THIS cacheContainer(String cacheContainer) {
             this.cacheContainer = cacheContainer;
-            return (THIS) this;
-        }
-
-        public final THIS jndiName(String jndiName) {
-            this.jndiName = jndiName;
-            return (THIS) this;
-        }
-
-        public final THIS module(String module) {
-            this.module = module;
-            return (THIS) this;
-        }
-
-        public final THIS start(String start) {
-            this.start = start;
             return (THIS) this;
         }