+++ testsuite/standalone/src/test/java/org/wildfly/extras/creaper/commands/infinispan/cache/AddDistributedCacheOnlineTest.java
@@ -2,14 +2,11 @@ package org.wildfly.extras.creaper.commands.infinispan.cache;
 
 import org.jboss.arquillian.junit.Arquillian;
 import org.junit.After;
-import org.junit.Assume;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.wildfly.extras.creaper.core.CommandFailedException;
 import org.wildfly.extras.creaper.core.ManagementClient;
-import org.wildfly.extras.creaper.core.ServerVersion;
 import org.wildfly.extras.creaper.core.online.ModelNodeResult;
 import org.wildfly.extras.creaper.core.online.OnlineManagementClient;
 import org.wildfly.extras.creaper.core.online.OnlineOptions;
@@ -34,16 +31,6 @@ public class AddDistributedCacheOnlineTest {
             .and("cache-container", "hibernate")
             .and("distributed-cache", TEST_CACHE_NAME);
 
-    @BeforeClass
-    public static void checkServerVersionIsSupported() throws Exception {
-        // check version is supported
-        ServerVersion serverVersion
-                = ManagementClient.online(OnlineOptions.standalone().localDefault().build()).version();
-        Assume.assumeFalse("The command is not compatible with WildFly 27 and above,"
-                        + " see https://github.com/wildfly-extras/creaper/issues/218.",
-                serverVersion.greaterThanOrEqualTo(ServerVersion.VERSION_20_0_0));
-    }
-
     @Before
     public void connect() throws Exception {
         client = ManagementClient.online(OnlineOptions.standalone().localDefault().build());
@@ -60,27 +47,21 @@ public class AddDistributedCacheOnlineTest {
     public void addCacheWithRequiredArgsOnly() throws CommandFailedException, IOException {
         AddDistributedCache cmd = new AddDistributedCache.Builder(TEST_CACHE_NAME)
                 .cacheContainer("hibernate")
-                .mode(CacheMode.SYNC)
                 .build();
         client.apply(cmd);
 
         ModelNodeResult resource = ops.readResource(TEST_CACHE_ADDRESS);
 
         assertTrue(resource.isSuccess());
-        assertEquals("SYNC", ops.readAttribute(TEST_CACHE_ADDRESS, "mode").stringValue());
     }
 
     @Test
     public void addCacheWithMoreArgs() throws CommandFailedException, IOException {
         AddDistributedCache cmd = new AddDistributedCache.Builder(TEST_CACHE_NAME)
                 .cacheContainer("hibernate")
-                .mode(CacheMode.SYNC)
-                .asyncMarshalling(true)
-                .queueFlushInterval(1234L)
                 .remoteTimeout(4321L)
                 .statisticsEnabled(false)
                 .capacityFactor(1.47)
-                .consistentHashStrategy(ConsistentHashStrategy.INTER_CACHE)
                 .l1lifespan(1780L)
                 .owners(3)
                 .segments(74)
@@ -90,16 +71,11 @@ public class AddDistributedCacheOnlineTest {
         ModelNodeResult resource = ops.readResource(TEST_CACHE_ADDRESS);
 
         assertTrue(resource.isSuccess());
-        assertEquals(CacheMode.SYNC.getMode(), ops.readAttribute(TEST_CACHE_ADDRESS, "mode").stringValue());
-        assertEquals(true, ops.readAttribute(TEST_CACHE_ADDRESS, "async-marshalling").booleanValue());
-        assertEquals(1234L, ops.readAttribute(TEST_CACHE_ADDRESS, "queue-flush-interval").longValue());
         assertEquals(4321L, ops.readAttribute(TEST_CACHE_ADDRESS, "remote-timeout").longValue());
         assertEquals(false, ops.readAttribute(TEST_CACHE_ADDRESS, "statistics-enabled").booleanValue());
         assertEquals(1780L, ops.readAttribute(TEST_CACHE_ADDRESS, "l1-lifespan").longValue());
         assertEquals(3, ops.readAttribute(TEST_CACHE_ADDRESS, "owners").intValue());
         assertEquals(74, ops.readAttribute(TEST_CACHE_ADDRESS, "segments").intValue());
         assertEquals(1.47, ops.readAttribute(TEST_CACHE_ADDRESS, "capacity-factor").doubleValue(), 0.001);
-        assertEquals(ConsistentHashStrategy.INTER_CACHE.getType(),
-                ops.readAttribute(TEST_CACHE_ADDRESS, "consistent-hash-strategy").stringValue());
     }
 }