+++ testsuite/standalone/src/test/java/org/wildfly/extras/creaper/commands/infinispan/cache/AddInvalidationCacheOnlineTest.java
@@ -2,14 +2,11 @@ package org.wildfly.extras.creaper.commands.infinispan.cache;
 
 import org.jboss.arquillian.junit.Arquillian;
 import org.junit.After;
-import org.junit.Assume;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.wildfly.extras.creaper.core.CommandFailedException;
 import org.wildfly.extras.creaper.core.ManagementClient;
-import org.wildfly.extras.creaper.core.ServerVersion;
 import org.wildfly.extras.creaper.core.online.ModelNodeResult;
 import org.wildfly.extras.creaper.core.online.OnlineManagementClient;
 import org.wildfly.extras.creaper.core.online.OnlineOptions;
@@ -34,16 +31,6 @@ public class AddInvalidationCacheOnlineTest {
             .and("cache-container", "hibernate")
             .and("invalidation-cache", TEST_CACHE_NAME);
 
-    @BeforeClass
-    public static void checkServerVersionIsSupported() throws Exception {
-        // check version is supported
-        ServerVersion serverVersion
-                = ManagementClient.online(OnlineOptions.standalone().localDefault().build()).version();
-        Assume.assumeFalse("The command is not compatible with WildFly 27 and above,"
-                        + " see https://github.com/wildfly-extras/creaper/issues/218.",
-                serverVersion.greaterThanOrEqualTo(ServerVersion.VERSION_20_0_0));
-    }
-
     @Before
     public void connect() throws Exception {
         client = ManagementClient.online(OnlineOptions.standalone().localDefault().build());
@@ -60,23 +47,18 @@ public class AddInvalidationCacheOnlineTest {
     public void addCacheWithRequiredArgsOnly() throws CommandFailedException, IOException {
         AddInvalidationCache cmd = new AddInvalidationCache.Builder(TEST_CACHE_NAME)
                 .cacheContainer("hibernate")
-                .mode(CacheMode.SYNC)
                 .build();
         client.apply(cmd);
 
         ModelNodeResult resource = ops.readResource(TEST_CACHE_ADDRESS);
 
         assertTrue(resource.isSuccess());
-        assertEquals("SYNC", ops.readAttribute(TEST_CACHE_ADDRESS, "mode").stringValue());
     }
 
     @Test
     public void addCacheWithMoreArgs() throws CommandFailedException, IOException {
         AddInvalidationCache cmd = new AddInvalidationCache.Builder(TEST_CACHE_NAME)
                 .cacheContainer("hibernate")
-                .mode(CacheMode.SYNC)
-                .asyncMarshalling(true)
-                .queueFlushInterval(1234L)
                 .remoteTimeout(4321L)
                 .statisticsEnabled(false)
                 .build();
@@ -85,9 +67,6 @@ public class AddInvalidationCacheOnlineTest {
         ModelNodeResult resource = ops.readResource(TEST_CACHE_ADDRESS);
 
         assertTrue(resource.isSuccess());
-        assertEquals(CacheMode.SYNC.getMode(), ops.readAttribute(TEST_CACHE_ADDRESS, "mode").stringValue());
-        assertEquals(true, ops.readAttribute(TEST_CACHE_ADDRESS, "async-marshalling").booleanValue());
-        assertEquals(1234L, ops.readAttribute(TEST_CACHE_ADDRESS, "queue-flush-interval").longValue());
         assertEquals(4321L, ops.readAttribute(TEST_CACHE_ADDRESS, "remote-timeout").longValue());
         assertEquals(false, ops.readAttribute(TEST_CACHE_ADDRESS, "statistics-enabled").booleanValue());
     }