+++ testsuite/standalone/src/test/java/org/wildfly/extras/creaper/commands/infinispan/cache/AddLocalCacheOnlineTest.java
@@ -2,14 +2,11 @@ package org.wildfly.extras.creaper.commands.infinispan.cache;
 
 import org.jboss.arquillian.junit.Arquillian;
 import org.junit.After;
-import org.junit.Assume;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.wildfly.extras.creaper.core.CommandFailedException;
 import org.wildfly.extras.creaper.core.ManagementClient;
-import org.wildfly.extras.creaper.core.ServerVersion;
 import org.wildfly.extras.creaper.core.online.ModelNodeResult;
 import org.wildfly.extras.creaper.core.online.OnlineManagementClient;
 import org.wildfly.extras.creaper.core.online.OnlineOptions;
@@ -33,16 +30,6 @@ public class AddLocalCacheOnlineTest {
     private static final Address TEST_CACHE_ADDRESS = Address.subsystem("infinispan")
             .and("cache-container", "hibernate")
             .and("local-cache", TEST_CACHE_NAME);
-
-    @BeforeClass
-    public static void checkServerVersionIsSupported() throws Exception {
-        // check version is supported
-        ServerVersion serverVersion
-                = ManagementClient.online(OnlineOptions.standalone().localDefault().build()).version();
-        Assume.assumeFalse("The command is not compatible with WildFly 27 and above,"
-                        + " see https://github.com/wildfly-extras/creaper/issues/218.",
-                serverVersion.greaterThanOrEqualTo(ServerVersion.VERSION_20_0_0));
-    }
 
     @Before
     public void connect() throws Exception {
@@ -70,21 +57,15 @@ public class AddLocalCacheOnlineTest {
 
     @Test
     public void addCacheWithMoreArgs() throws CommandFailedException, IOException {
-        String jndiName = "java:/MyAwesomeCache";
-        String module = "org.hibernate.infinispan";
         AddLocalCache cmd = new AddLocalCache.Builder(TEST_CACHE_NAME)
                 .cacheContainer("hibernate")
                 .statisticsEnabled(false)
-                .jndiName(jndiName)
-                .module(module)
                 .build();
         client.apply(cmd);
 
         ModelNodeResult resource = ops.readResource(TEST_CACHE_ADDRESS);
 
         assertTrue(resource.isSuccess());
-        assertEquals(jndiName, ops.readAttribute(TEST_CACHE_ADDRESS, "jndi-name").stringValue());
-        assertEquals(module, ops.readAttribute(TEST_CACHE_ADDRESS, "module").stringValue());
         assertEquals(false, ops.readAttribute(TEST_CACHE_ADDRESS, "statistics-enabled").booleanValue());
     }
 }